Send and receive rigid body optimization messages. Remove unnecessary print of user's XML on login. Correct module load order since TundraLogic now depends on physics. Use float for frame deltatime consistently in TundraLogic & BulletPhysics modules.

+++ src/Plugins/BulletPhysics/PhysicsWorld.cpp
@@ -417,7 +417,7 @@ btDiscreteDynamicsWorld* PhysicsWorld::BulletWorld() const
     return impl->world;
 }
 
-void PhysicsWorld::Simulate(f64 frametime)
+void PhysicsWorld::Simulate(float frametime)
 {
     if (!runPhysics_)
         return;
